backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
gunicorn==21.2.0
sqlalchemy==2.0.28
pydantic==2.6.3
pydantic-settings==2.1.0
pydantic[email]==2.6.3
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
python-multipart==0.0.7
python-dotenv==1.0.1
greenlet==3.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
google-cloud-secret-manager==2.19.0